Room (#104)
Sorry this PR is so big. I had to duplicate a lot of classes, add a
bunch of type converters and mapping classes, the Dao's, etc.

I'm using the [one-to-many
relationship](https://developer.android.com/training/data-storage/room/relationships#one-to-many)
for home timeline. If you run the app and use the app inspector you can
easily view the tables. The main table is `homeTimeline`. It has 6
columns; `statusId`, `createdAt`, `accountId`, `pollId`,
`boostedStatusId`, `boostedStatusAccountId`, and `boostedPollId`. Then
there is an object we can query called `HomeTimelineStatusWrapper`. This
object builds the relationship between the objects in `homeTimeline` and
the objects in `accounts`, `statuses`, and `polls` tables, and it holds
the actual objects for accounts and statuses.

I can explain it better in person. Here is a miro board that diagrams
how paging works in general https://miro.com/app/board/uXjVMt6i4ss=/

---------

Co-authored-by: John Oberhauser <j.git-global@obez.io>