block and mute (#118)
Adding functionality to the block and mute buttons in the status card
overflow menu.

Removing the follow button in the status card overflow menu. I want to
keep follow just in the account page for now because the account object
returned by mastodon doesn't include the following state, so extra
network calls would be required. We could possibly add a follow option
in the status card in the future, but I don't think it's necessary at
the moment.

Co-authored-by: John Oberhauser <j.git-global@obez.io>