Simplifying colors (#189)
@devotaaabel I think you were just copying what firefox did for their
theme colors, right? It seems a little complex to me, and I'm trying to
figure out if there is any benefits to it. When I was adding some colors
to the other day, I found it tedious having to update the constructor,
the variables, the update method, and the copy method. Any reason we
can't simplify it like so?

---------

Co-authored-by: John Oberhauser <j.git-global@obez.io>