Making videos / gifs not "pop in" (#197)
When scrolling up in a timeline, videos and gifs will "pop in" and push
content down. It's very jarring. It does this because as the video
player loads, the height gets set. Before it loads the height is 0. This
can be fixed by setting a static height for videos. But what should the
height be? If we hardcode a height, the aspect ratio could be different
on different devices. So instead, I'm just setting the aspect ratio
directly.

- Making videos have an aspect ratio of 1:1
- Giving videos a rounded corner shape
- Cleaning up the video player file just a tad

Co-authored-by: John Oberhauser <j.git-global@obez.io>