Retain scroll (#193)
When navigating back to a list, the `lazyPagingItems` seem to have a
list size of 0 for a split second before going back to where it was.
This causes the list scroll state to reset at 0, losing the scroll
position. I'm creating a second scroll state to use when the list is not
0... I'm not super confident in this solution, but it does work. There
is probably a better way though. Ideally, the flow would not emit an
empty list initially. I'm not sure we can prevent it though.

- Retaining scroll position by providing two scroll states
- Updating top level navigation so a new feed isn't created every time
we navigate to the feed

---------

Co-authored-by: John Oberhauser <j.git-global@obez.io>