fixing timelines (#236)
For some reason, I started getting 19 statuses back from the API during
home timeline paging, even though I'm requesting 20. Our current logic
is that if the response count != the request page size, that means we
are done paging. We should instead be checking if there is a `next` link
header.

- Updating all timelines to grab to look at the link header for
determining if we should continue paging

Co-authored-by: John Oberhauser <j.git-global@obez.io>