Adding coroutine scope that cancels on logout (#253)
Useful reading for understanding this

https://medium.com/androiddevelopers/coroutines-patterns-for-work-that-shouldnt-be-cancelled-e26c40f142ad

https://developer.android.com/kotlin/coroutines/coroutines-best-practices#create-coroutines-data-layer

- Creating a singleton `appScope` that is a coroutine scope with a
`SupervisorJob`
- The `SupervisorJob` gets canceled on logout, and a new `SupervisorJob`
takes it's place
- Moving a bunch of repository function to use cases
- For functions in the repository that we don't want to be canceled when
the `viewModelScope` is canceled, we make the following changes
- Within the dispatcher context, we launch a coroutine from `appScope`
with `async` + `await`. Using `async` + `await` allows the
`viewmodelScope` to cancel immediately when the user navigates away from
the screen. If we use `launch` + `join`, then `viewmodelScope` doesn't
cancel until after the `appScope` work is done. Same with just using
`withContext(appScope)`
- The appscope coroutine context is set to `dispatcherIo` which is an
injected dispatcher that defaults to `Dispatchers.IO`.
- The work is wrapped in a try catch that will rethrow any exception,
but wrap it in a custom exception. We specify the exception like this so
we can catch the specific exception in the viewmodel. We need to move
away from catching `Exception` for these because that will also catch a
`CancellationException`. We don't want to catch a
`CancellationException` because we don't want error toasts showing up
just because the user navigated away from the screen. A
`CancellationException` will not crash the app if not caught.
- Updating unit tests

---------

Co-authored-by: John Oberhauser <j.git-global@obez.io>