Setup polls with data source/repository pattern (#274)
I wanted to start with a simple example before I apply the pattern to
the rest of the app. I can do the rest of the daos in more PRs (it might
take a while, a few seem like rabbit holes).

Basically I started by making the poll dao internal in the database
module, and adding a `PollLocalDataSource` class as the module's public
API. Everything else should be internal, and that way we'll have an
enforcement to keep the layers.

I created a new `PollRepository`, and I think we can make repositories
for any single type of data. They can include either local or remote
data sources (or both), and you won't really know which until you look
at it. It should be pretty clear, as a repository should only really
have two parameters (a `localDataSource` and/or a `remoteDataSource`).
The idea is to create use cases for anything that combines multiple
streams of data.

I've also been moving single-use mapper extension functions to their
usage sites and making them private. The `model` module is supposed to
be the language between our modules, so mappers like this should be
private or internal where it makes sense.

I think using interfaces at this point would be premature. Let's wait
until we are actually in a situation that warrants it.