Code cleanup part 2 (#295)
This is part two of code cleanup.

I deleted all our detekt baselines, and fixed things manually (and
deleted some unused things). I think baseline files are good for large
projects that would be unreasonable to fix all problems for, but I just
fixed them all in an hour so I think it was worth it.

I also re-added the `NewLineAtEndOfFile` rule. We'll see how it goes and
possibly re-add it if it gets annoying.