Removing broken field from account (#349)
This change will fix this issue
https://github.com/MozillaSocial/mozilla-social-android/issues/348

The mastodon API spec says the `last_status_at` field in an `Account`
body should be in a format like this `2019-11-24T15:49:42.251Z`, however
it's coming back like this `2019-11-24`. We were parsing it the way
mastodon gives it to us, but it's incorrect according to the docs.
Friendica sends back the correct format, which is why Friendica sign in
wasn't working. I'm just removing the field because it's broken in
mastodon and we aren't using it anyways.

---------

Co-authored-by: John Oberhauser <j.git-global@obez.io>