Periodically purge database (#353)
Adding a work manager job that purges the database of most entries every
20 days.

Something to note: if the app is open (or backgrounded) while the job
runs, it can put the app in a weird state. To fix this, I'm observing
the job state and restarting the app when it's complete. The job won't
run unless the user's phone is idle, so they shouldn't notice this.

I also added a new screen in settings called "Developer options" that
only shows up in debug builds. In here I put a button you can click to
test the work manager job.

---------

Co-authored-by: John Oberhauser <j.git-global@obez.io>