Setup baseline profile (#545)
I'm setting up a very basic baseline profile. It's a lot of work to get
it to run through common code paths due to login. To do this the right
way, I would need to set up mocked network calls, change how login works
so I can provide a mock login usecase that doesn't send the user to a
custom tab, get the baseline profile tasks to run in CI (which I think
was an issue with gradle managed emulator devices), and figure out how
to provide mocks when the baseline profile module doesn't depend on the
app. It's all kinda confusing and I don't have time to figure it all
out. In the meantime I'm just adding a profile that ran through an
already logged in session, without navigating anywhere. Hopefully it's
enough to help with any scrolling issues in the near term.

---------

Co-authored-by: John Oberhauser <j.git-global@obez.io>